refactor: 폴더구조 MVC로 변경
