+++ src/main/java/step2_2/domain/InvalidMoneyInputException.java
@@ -1,4 +1,4 @@
-package step2_2;
+package step2_2.domain;
 
 public class InvalidMoneyInputException extends RuntimeException {
