+++ src/main/java/step2_2/domain/Lotto.java
@@ -1,4 +1,4 @@
-package step2_2;
+package step2_2.domain;
 
 import static java.util.stream.Collectors.toList;
